+++ nectar/src/bitcoin/fee.rs
@@ -5,54 +5,43 @@ const ESTIMATE_FEE_TARGET: u32 = 3;
 
 #[derive(Clone, Debug)]
 pub struct Fee {
-    config: config::BitcoinFeeStrategy,
-    max_fee: bitcoin::Amount,
+    config: config::BitcoinFees,
     client: bitcoin::Client,
 }
 
 impl Fee {
     // TODO: Improve this API, the client is not needed
     // if we use the static fee
-    pub fn new(
-        config: config::BitcoinFeeStrategy,
-        max_btc_fee: bitcoin::Amount,
-        client: bitcoin::Client,
-    ) -> Self {
-        Self {
-            config,
-            max_fee: max_btc_fee,
-            client,
-        }
+    pub fn new(config: config::BitcoinFees, client: bitcoin::Client) -> Self {
+        Self { config, client }
     }
 
-    pub async fn kbyte_rate(&self) -> Result<Amount> {
-        let byte_rate = self.byte_rate().await?;
-        byte_rate
-            .checked_mul(1000)
-            .context("Could not mul byte rate")
+    pub async fn kvbyte_rate(&self) -> Result<Amount> {
+        let rate = self.vbyte_rate().await?;
+        rate.checked_mul(1000).context("Could not mul byte rate")
     }
 
-    pub async fn byte_rate(&self) -> Result<Amount> {
-        use crate::config::BitcoinFeeStrategy::*;
+    pub async fn vbyte_rate(&self) -> Result<Amount> {
+        use crate::config::BitcoinFees::*;
         match self.config {
-            SatsPerByte(amount) => Ok(amount),
-            BitcoindEstimateSmartfee(mode) => {
-                let kbyte_rate = self
+            SatsPerByte(fee) => Ok(fee),
+            BitcoindEstimateSmartfee { mode, .. } => {
+                let kvbyte_rate = self
                     .client
                     .estimate_smart_fee(ESTIMATE_FEE_TARGET, Some(mode.into()))
                     .await
                     .map(|res| res.kbyte_rate)?;
 
                 // Return rate per byte
-                kbyte_rate
+                kvbyte_rate
                     .checked_div(1000)
                     .context("Could not div kbyte rate")
             }
         }
     }
 
-    pub fn max_fee(&self) -> bitcoin::Amount {
-        self.max_fee
+    pub fn max_tx_fee(&self) -> bitcoin::Amount {
+        self.config.max_tx_fee()
     }
 }
 
@@ -61,7 +50,6 @@ impl crate::StaticStub for Fee {
     fn static_stub() -> Self {
         Self {
             config: Default::default(),
-            max_fee: bitcoin::Amount::ZERO,
             client: bitcoin::Client::new("http://example.com/".parse().unwrap()), // Not used
         }
     }