+++ nectar/src/main.rs
@@ -135,8 +135,7 @@ async fn main() -> Result<()> {
             println!("{}", deposit);
         }
         Command::Withdraw(arguments) => {
-            let ethereum_gas_price =
-                ethereum::GasPrice::new(settings.maker.fee_strategies.ethereum);
+            let ethereum_gas_price = ethereum::GasPrice::new(settings.ethereum.gas_price);
             let tx_id = withdraw(
                 ethereum_wallet.expect("could not initialise ethereum wallet"),
                 ethereum_gas_price,
@@ -150,14 +149,9 @@ async fn main() -> Result<()> {
         Command::DumpConfig => unreachable!(),
         Command::ResumeOnly => {
             let bitcoind_client = bitcoin::Client::new(settings.bitcoin.bitcoind.node_url.clone());
-            let bitcoin_fee = bitcoin::Fee::new(
-                settings.maker.fee_strategies.bitcoin,
-                settings.maker.maximum_possible_fee.bitcoin,
-                bitcoind_client,
-            );
+            let bitcoin_fee = bitcoin::Fee::new(settings.bitcoin.fees, bitcoind_client);
 
-            let ethereum_gas_price =
-                ethereum::GasPrice::new(settings.maker.fee_strategies.ethereum.clone());
+            let ethereum_gas_price = ethereum::GasPrice::new(settings.ethereum.gas_price.clone());
 
             resume_only(
                 settings,